Move password decryption to background.js
Fixes the issue with password entry failing when the GPG keychain needs
unlocking to decrypt the password.

When the GPG keychain is locked, the pinentry program will steal focus from
Chrome. This will cause the extension popup to close.

Previously all of our interaction with pass was happening in popup.js, which
meant that when the popup was closed by Chrome losing focus, the JavaScript
would stop running. Once the keychain was unlocked, we'd get a response, but
there wouldn't be anything listening for it anymore.

Doing the final decryption step via a background script instead of a popup
script means we can still listen for the response after the popup has
closed.